requirements.txt
django
djangorestframework
djangorestframework_simplejwt
pillow
django-celery
django-celery-email